make the send methods return a size_t so we don't do signed/unsigned compares
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@16333 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/BufferedNetworkMessage.h
@@ -72,8 +72,8 @@ class NetworkMessageTransferCallback
 public:
   virtual ~NetworkMessageTransferCallback(){};
 
-  virtual int send ( NetHandler *handler, void *data, size_t size ) = 0;
-  virtual int broadcast ( void *data, size_t size, int mask, int code  ) = 0;
+  virtual size_t send ( NetHandler *handler, void *data, size_t size ) = 0;
+  virtual size_t broadcast ( void *data, size_t size, int mask, int code  ) = 0;
 };
 
 class BufferedNetworkMessageManager : public Singleton<BufferedNetworkMessageManager>
